refactor(scripts): split post-review.mjs and validate-findings.mjs (#3795)
Both files had grown past the house 400-line module-size budget through a
run of review-lane fixes, each raising the allowlist. Split along the seams
the code already has, mechanically -- no behaviour change:

post-review.mjs -> scripts/review/lib/
  - review-findings.mjs: findings.json reading and the marker format
  - review-comments.mjs: posting/reading back GitHub comments
  - review-summary.mjs: the human-facing summary/contradiction text
  - post-review-error.mjs: the shared error class (avoids a circular import)

validate-findings.mjs -> scripts/review/lib/
  - finding-schema.mjs: schema/anchor validation (validate/checkSchema/
    validateFindings/omittedForPromptPaths)
  - finding-proof-of-work.mjs: checkProofOfWork/siblingVerifies
  - finding-sanitizers.mjs: sanitizeBody/sanitizeLabel/sanitizePath
  - review-input-reader.mjs: reading the model's raw output and our own
    review-input JSON
  - validate-findings-error.mjs: the shared error class

Every existing export of both files is re-exported unchanged, so no
consumer (rubric-eval.mjs, run-judge.mjs, lane-canary.test.mjs, etc.)
needed to change its imports.

validate-findings.test.mjs's REASONS-coverage inventory test scanned only
validate-findings.mjs's own source text for ValidateFindingsError raise
sites; since the requested split necessarily moves several raise sites
(checkSchema, checkProofOfWork, validate, readInput) into sibling files,
its source scan now reads all four files the raise sites can live in. The
assertion logic is unchanged -- only what counts as "this module" widened
to match the new file boundary.

Lowered scripts/module-size-allowlist.txt: both files dropped under 400
lines, so their allowlist rows are deleted rather than lowered.

Closes #3795
