+++ scripts/module-size-allowlist.txt
@@ -387,10 +387,8 @@
    980 scripts/moonshot/m5-constrained-decoding/tasks-tier2.mjs
    680 scripts/review/build-context-pack.mjs
    679 scripts/review/build-review-input.mjs
-  1062 scripts/review/post-review.mjs
    443 scripts/review/rubric-eval.mjs
    551 scripts/review/run-reviewer.mjs
-  1103 scripts/review/validate-findings.mjs
    688 scripts/source-text-assertion-detect.mjs
    504 scripts/test-geometry-regression.mjs
   2196 scripts/test-wasm-contract.mjs
