+++ scripts/review/lib/validate-findings-error.mjs
@@ -1,0 +1,17 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/**
+ * The one error class `validate-findings.mjs` and its split-out siblings all
+ * throw. Its own file (module-size budget, #3795) so every sibling can throw
+ * it without importing back through `validate-findings.mjs` -- which would be
+ * circular, since that file imports `validate()` from `lib/finding-schema.mjs`.
+ */
+
+export class ValidateFindingsError extends Error {
+  constructor(reason, message) {
+    super(message);
+    this.reason = reason;
+  }
+}
