+++ scripts/review/validate-findings.mjs
@@ -147,40 +147,32 @@
  *      catches the throttled case only when it also stopped quoting.
  */
 
-import { readFileSync, writeFileSync, rmSync } from 'node:fs';
-import { createHash } from 'node:crypto';
+import { writeFileSync, rmSync } from 'node:fs';
 import { isMainEntry } from '../lib/is-main-entry.mjs';
+import { ValidateFindingsError } from './lib/validate-findings-error.mjs';
 // quotableLines/quoteAppearsIn/lineIsAdded/addedLinesMatching moved to
-// ./quote-line-coupling.mjs (module-size budget). Imported (not `export ...
-// from`) because this file also calls them itself, and re-exported below so
-// every existing import of them from this file keeps working unchanged.
+// ./quote-line-coupling.mjs (module-size budget). Re-exported below so every
+// existing import of them from this file keeps working unchanged.
 import { quotableLines, quoteAppearsIn, lineIsAdded, addedLinesMatching } from './quote-line-coupling.mjs';
-// One constant, imported rather than re-spelled: a reworded copy here would
-// stop matching the rows build-review-input writes, and the partial-review
-// marker would silently claim a full review (#3679).
-import { isUnread } from './build-review-input.mjs';
-
-/**
- * The terminal sentinel the prompt requires as the LAST field. Its whole job is to
- * be absent when the response stopped early, so it is compared with `===` against
- * a literal and never matched loosely -- a `startsWith` here would accept
- * `ifc-lite-review-v1-partial` and defeat the check it exists to be.
- */
-export const SENTINEL = 'ifc-lite-review-v1';
-
-/**
- * What may survive VALIDATION. It was 5, which made sense when the reviewer was
- * the last line of defence and every finding it wrote went straight onto the PR.
- * With a judge downstream, capping here would throw candidates away before
- * anything could weigh them -- precision enforced at generation time, which is
- * what produced a 2% finding rate.
- *
- * The cap on what reaches a HUMAN is a different question and lives in
- * post-review.mjs, the only module on the posting path that always runs. It was
- * briefly declared here too; two constants of the same name and value in two
- * modules agree only until someone edits one.
- */
-export const MAX_FINDINGS = 12;
+// readText/stripFence/parseRaw/readInput moved to ./lib/review-input-reader.mjs
+// (module-size budget, #3795). Imported (not `export ... from`) because
+// main() below calls `readInput` and `parseRaw` itself, and re-exported so
+// every existing import of `stripFence`/`readInput` from this file keeps
+// working unchanged.
+import { readText, stripFence, parseRaw, readInput } from './lib/review-input-reader.mjs';
+// SENTINEL/MAX_FINDINGS/validate/omittedForPromptPaths moved to
+// ./lib/finding-schema.mjs (module-size budget, #3795). Imported (not `export
+// ... from`) because main() below calls `validate` and `omittedForPromptPaths`
+// itself, and re-exported so every existing import of these four names from
+// this file keeps working unchanged.
+import { SENTINEL, MAX_FINDINGS, validate, omittedForPromptPaths } from './lib/finding-schema.mjs';
+// MAX_BODY_CHARS/sanitizeBody/sanitizeLabel/sanitizePath moved to
+// ./lib/finding-sanitizers.mjs (module-size budget, #3795). Re-exported below.
+import { MAX_BODY_CHARS, sanitizeBody, sanitizeLabel, sanitizePath } from './lib/finding-sanitizers.mjs';
+// siblingVerifies moved to ./lib/finding-proof-of-work.mjs (module-size
+// budget, #3795), alongside checkProofOfWork, which stays private there.
+// Re-exported below.
+import { siblingVerifies } from './lib/finding-proof-of-work.mjs';
 
 /**
  * Every reason this module can exit with.
@@ -212,75 +204,6 @@ export const REASONS = new Set([
   'OUT_UNWRITABLE',
 ]);
 
-/** GitHub renders long comments fine; a reviewer reading twenty of them does not. */
-export const MAX_BODY_CHARS = 1500;
-const TRUNCATION_NOTE = '\n\n[truncated by validate-findings]';
-
-/**
- * A quote has to be long enough to BE evidence. `}` appears in every patch ever
- * written and quoting it proves nothing, so a proof-of-work quote that short is
- * indistinguishable from a guess.
- *
- * The two bounds differ on purpose. The riskiest-change quote is the ONE piece of
- * evidence standing between us and #1644, it is fatal when it fails, and the
- * prompt asks for a substantive line -- so it is held to eight characters. A
- * per-finding quote is backed by a second, independent check (the line must fall
- * inside an added range) and its failure DROPS a possibly-real finding, so it is
- * held to three: enough to exclude the empty and one-character cases that match
- * everything, not so much that a finding about `x = 0;` is thrown away.
- */
-const MIN_PROOF_QUOTE_CHARS = 8;
-const MIN_FINDING_QUOTE_CHARS = 3;
-
-/** A class label is a short tag, not a place to smuggle a paragraph. */
-const MAX_CLASS_CHARS = 60;
-
-/**
- * THE TOKEN THAT MUST NOT SURVIVE INTO A POSTED BODY.
- *
- * Matched case-insensitively even though `check-review-posted.mjs`'s MARKER_RE is
- * case-sensitive: defanging more than the gate matches is free, and the reverse
- * mistake is a hole. The replacement swaps the SECOND ASCII hyphen for U+2011
- * NON-BREAKING HYPHEN, which reads identically to a human and cannot match a
- * pattern that requires `-`. A zero-width space would work equally well and be
- * invisible; a visible-but-inert token is preferred so a reader looking at a
- * posted comment can SEE that something was defanged rather than wonder why the
- * gate ignored it.
- */
-const MARKER_TOKEN_RE = /ifc-lite-review/gi;
-/**
- * A REPLACER, not a fixed string, and that is the correction. The pattern is
- * case-INSENSITIVE while the replacement was a lowercase literal, so defanging
- * REWROTE the text it was defanging: `docs/IFC-Lite-Review-Lane.md` came out as
- * `docs/ifc-lite‑review-Lane.md`, a name that exists nowhere. That is the same
- * class as the 60-char `class` cap this file already records for rewriting real
- * paths into names that exist nowhere, arriving by a different door -- and it
- * lands on an advisory list whose whole job is naming files a human then reads.
- *
- * Swaps the SECOND ASCII hyphen of whatever was matched for U+2011 and touches
- * nothing else, so case survives.
- */
-const defangToken = (match) => `${match.slice(0, 8)}\u2011${match.slice(9)}`;
-
-/** Whole HTML comments, non-greedy, including multi-line ones. */
-const HTML_COMMENT_RE = /<!--[\s\S]*?-->/g;
-
-/**
- * A dangling `<!--` left after the pass above (an UNCLOSED comment). It cannot
- * carry a marker on its own -- the gate's pattern needs the closing `-->` -- but
- * it can swallow whatever the poster appends after it when GitHub renders the
- * comment, including the real marker. Neutralised rather than deleted so the text
- * a human wrote is still legible.
- */
-const DANGLING_COMMENT_OPEN_RE = /<!--/g;
-
-export class ValidateFindingsError extends Error {
-  constructor(reason, message) {
-    super(message);
-    this.reason = reason;
-  }
-}
-
 /**
  * A Map, not an object literal, for the reason check-review-posted.mjs records:
  * `{...}[name]` reaches Object.prototype, so `--constructor x` returns a truthy
@@ -306,720 +229,14 @@ export function parseArgs(argv) {
   return out;
 }
 
-/** @param {string} path @param {string} kind */
-function readText(path, kind) {
-  try {
-    return readFileSync(path, 'utf8');
-  } catch (err) {
-    throw new ValidateFindingsError(
-      kind === 'raw' ? 'RAW_UNREADABLE' : 'INPUT_UNREADABLE',
-      `Cannot read \`${path}\`: ${err.code || err.message}. ` +
-        (kind === 'raw'
-          ? 'A MISSING model output is not an empty clean review; it is the #1644 shape at its most ' +
-            'extreme. REMEDY: read the review step\'s log -- it ran and produced no file.'
-          : 'REMEDY: fix the step that builds review-input.json.'),
-    );
-  }
-}
-
-/**
- * Strip ONE leading fence and its matching trailing fence. Nothing else is
- * repaired: see hole 3. A leading fence with no closing one is left alone, which
- * makes the remainder fail to parse if it is genuinely truncated and parse if it
- * is not -- the honest outcome either way.
- *
- * @param {string} text
- */
-export function stripFence(text) {
-  const t = String(text).trim();
-  if (!t.startsWith('```')) return t;
-  const nl = t.indexOf('\n');
-  if (nl === -1) return t;
-  // ``` optionally followed by a bare language tag, and NOTHING else. `~~~json {`
-  // or ```json trailing junk is not a fence this will strip, because stripping a
-  // line it does not understand is a repair.
-  if (!/^```[A-Za-z0-9_+-]*$/.test(t.slice(0, nl).trim())) return t;
-  let body = t.slice(nl + 1);
-  const close = body.lastIndexOf('```');
-  if (close !== -1 && body.slice(close + 3).trim() === '') body = body.slice(0, close);
-  return body.trim();
-}
-
-/**
- * The model's text to an object, or a classified refusal.
- *
- * The plain-object check is NOT folded into the schema pass below and runs before
- * the sentinel check, because it is a precondition of both: reaching for `.end` on
- * `null` throws a TypeError past this file's catch and prints a stack trace where
- * a remedy should be. `[1,2]` and `"done"` are refused for the same reason. That
- * is SCHEMA_INVALID rather than RESPONSE_TRUNCATED on purpose -- a response of the
- * wrong SHAPE is a prompt problem, not a length problem, and the remedies differ.
- *
- * @param {string} text
- */
-export function parseRaw(text) {
-  const stripped = stripFence(text);
-  if (stripped === '') {
-    throw new ValidateFindingsError(
-      'RAW_EMPTY',
-      'The model produced no output at all. This is the #1644 silent no-op: the step exits 0 having ' +
-        'reviewed nothing. It is NOT a clean review. REMEDY: re-run, and read `num_turns` in the ' +
-        'review step\'s log if it recurs.',
-    );
-  }
-  let parsed;
-  try {
-    parsed = JSON.parse(stripped);
-  } catch (err) {
-    throw new ValidateFindingsError(
-      'RAW_UNPARSEABLE',
-      `The model's output is not JSON: ${err.message}. One leading/trailing \`\`\` fence is stripped ` +
-        'and nothing else is repaired, deliberately. REMEDY: tighten the prompt\'s output instruction. ' +
-        'Do not add a repair pass here -- a repairer that guesses is a second unreviewed model.',
-    );
-  }
-  if (parsed === null || typeof parsed !== 'object' || Array.isArray(parsed)) {
-    throw new ValidateFindingsError(
-      'SCHEMA_INVALID',
-      `The model's output parsed as ${parsed === null ? 'null' : Array.isArray(parsed) ? 'an array' : typeof parsed}, ` +
-        'not an object. Reaching for a field on it would throw past this file\'s catch and print a ' +
-        'stack trace instead of a remedy. REMEDY: fix the prompt.',
-    );
-  }
-  return parsed;
-}
-
-/** @param {unknown} v */
-const isNonEmptyString = (v) => typeof v === 'string' && v.trim() !== '';
-
-/**
- * The review-input WE built. Validated as strictly as the model's output, because
- * a broken input makes every check below pass vacuously -- and a vacuous pass here
- * is a green tick over an unreviewed diff, which is the entire failure this lane
- * exists to close.
- *
- * @param {string} path
- */
-export function readInput(path) {
-  const raw = readText(path, 'input');
-  let cfg;
-  try {
-    cfg = JSON.parse(raw);
-  } catch (err) {
-    throw new ValidateFindingsError('INPUT_INVALID', `\`${path}\` is not valid JSON: ${err.message}`);
-  }
-  if (cfg === null || typeof cfg !== 'object' || Array.isArray(cfg)) {
-    throw new ValidateFindingsError('INPUT_INVALID', `\`${path}\` must be a JSON object.`);
-  }
-  if (typeof cfg.headSha !== 'string' || !/^[0-9a-f]{40}$/.test(cfg.headSha)) {
-    throw new ValidateFindingsError(
-      'INPUT_INVALID',
-      `\`headSha\` must be a full 40-hex commit; got ${JSON.stringify(cfg.headSha)}. It is copied ` +
-        'verbatim into findings.json so the marker names a commit the MODEL never chose.',
-    );
-  }
-  if (!Array.isArray(cfg.files) || cfg.files.length === 0) {
-    throw new ValidateFindingsError(
-      'INPUT_INVALID',
-      '`files` must be a non-empty array. With zero files every check below passes having verified ' +
-        'nothing, which is a scan of nothing reported as a clean scan (#3194). REMEDY: fix the step ' +
-        'that builds review-input.json, or skip the review lane entirely for an empty diff.',
-    );
-  }
-  const files = new Map();
-  for (const [i, f] of cfg.files.entries()) {
-    if (f === null || typeof f !== 'object' || Array.isArray(f)) {
-      throw new ValidateFindingsError('INPUT_INVALID', `\`files[${i}]\` is not an object.`);
-    }
-    if (!isNonEmptyString(f.path)) {
-      throw new ValidateFindingsError('INPUT_INVALID', `\`files[${i}].path\` must be a non-empty string.`);
-    }
-    if (typeof f.patch !== 'string') {
-      throw new ValidateFindingsError('INPUT_INVALID', `\`files[${i}].patch\` must be a string.`);
-    }
-    if (files.has(f.path)) {
-      // Set equality still passes with a duplicate, and "that file's patch"
-      // silently becomes "whichever copy won" -- a check that reads as precise
-      // while adjudicating an arbitrary half of the input.
-      throw new ValidateFindingsError(
-        'INPUT_INVALID',
-        `\`${f.path}\` appears twice in \`files\`. Which patch a finding is checked against would be ` +
-          'decided by array order. REMEDY: de-duplicate in the builder.',
-      );
-    }
-    const ranges = f.addedLineRanges;
-    if (!Array.isArray(ranges)) {
-      throw new ValidateFindingsError('INPUT_INVALID', `\`files[${i}].addedLineRanges\` must be an array.`);
-    }
-    for (const [j, r] of ranges.entries()) {
-      const bad =
-        !Array.isArray(r) ||
-        r.length !== 2 ||
-        !Number.isInteger(r[0]) ||
-        !Number.isInteger(r[1]) ||
-        r[0] < 1 ||
-        r[1] < r[0];
-      if (bad) {
-        throw new ValidateFindingsError(
-          'INPUT_INVALID',
-          `\`files[${i}].addedLineRanges[${j}]\` must be [start, end] integers with 1 <= start <= end; ` +
-            `got ${JSON.stringify(r)}.`,
-        );
-      }
-    }
-    files.set(f.path, { path: f.path, patch: f.patch, addedLineRanges: ranges });
-  }
-  const unreviewable = cfg.unreviewable === undefined ? [] : cfg.unreviewable;
-  if (
-    !Array.isArray(unreviewable) ||
-    unreviewable.some((u) => !u || typeof u !== 'object' || !isNonEmptyString(u.path))
-  ) {
-    throw new ValidateFindingsError(
-      'INPUT_INVALID',
-      '`unreviewable` must be an array of objects each carrying a non-empty `path`. Annotated STRINGS were the ' +
-        'earlier shape and made the overlap check below unable to match anything -- an inert ' +
-        'guard that reads as a live one.',
-    );
-  }
-  for (const { path: p } of unreviewable) {
-    if (files.has(p)) {
-      throw new ValidateFindingsError(
-        'INPUT_INVALID',
-        `\`${p}\` is listed in BOTH \`files\` and \`unreviewable\`. The model is told those two lists ` +
-          'mean opposite things, so proof of work would demand it both review and not review the file.',
-      );
-    }
-  }
-  // `contextPack` IS CARRIED. Without it `input.contextPack` is undefined at the
-  // siblingVerifies call, that helper takes its "no sibling excerpts were
-  // provided" branch every single time, and EVERY finding naming a sibling is
-  // dropped as fabricated -- including ones whose sibling is real and was in the
-  // pack the reviewer was shown. A review whose findings all name siblings then
-  // dies as VALIDATION_EMPTY: red job, no marker, and no re-run can clear it.
-  // The check was written, tested by hand, and wired to a value that never
-  // arrived.
-  return { headSha: cfg.headSha, files, unreviewable, contextPack: cfg.contextPack ?? null };
-}
 
 export { quotableLines, quoteAppearsIn, lineIsAdded, addedLinesMatching };
 
-/**
- * THE SECURITY BOUNDARY. Everything the model controls that can reach a posted
- * comment body goes through here.
- *
- * Order is deliberate and each step depends on the one before it:
- *
- *   1. Whole HTML comments are REMOVED. They can carry a forged marker, and they
- *      are invisible in the rendered comment, so anything hiding in one is hiding
- *      on purpose.
- *   2. Any remaining `<!--` -- an unclosed comment -- is broken, because it would
- *      otherwise swallow the real marker the poster appends after this body.
- *   3. The literal token `ifc-lite-review` is broken EVERYWHERE, not only inside
- *      comments.
- *
- *      WHAT WAS ACTUALLY MEASURED, because the obvious claim here is wrong.
- *      Mutation-testing this file showed that against the gate's CURRENT
- *      MARKER_RE, steps 1 and 2 are already sufficient on their own: that pattern
- *      requires a literal `<!--`, and after those two steps no `<!--` survives in
- *      the output at all. So the three steps are mutually redundant there, and it
- *      would be false to call this one "the" defence.
- *
- *      It earns its place on the two cases the others do not cover. First, the
- *      token appears in ORDINARY TEXT that is not in a comment -- this lane's own
- *      source carries it (MARKER_RE in check-review-posted.mjs, and this file), so
- *      a model reviewing that diff quotes it and a reviewer writing about it types
- *      it. Second, it is the only step that still holds if the gate's pattern is
- *      ever loosened to match the token outside an HTML comment, which is a change
- *      a future editor could make in check-review-posted.mjs without ever reading
- *      this file.
- *   4. `@` before a word character gets a zero-width space, so a body cannot
- *      summon a person or a team into a thread. (An email address in a body picks
- *      up the same treatment. That is a cosmetic cost on a rare input, taken
- *      knowingly rather than adding a cleverer pattern with a hole in it.)
- *   5. The length cap runs LAST, so the final string is genuinely within the cap:
- *      steps 1-4 change the length in both directions, and capping before them
- *      would let defanging push the result back over. Truncation can only DELETE
- *      trailing text, so it cannot construct a marker out of what remains.
- *
- * @param {unknown} text
- */
-export function sanitizeBody(text) {
-  let out = defangDangerous(text);
-  if (out.length > MAX_BODY_CHARS) {
-    out = out.slice(0, MAX_BODY_CHARS - TRUNCATION_NOTE.length) + TRUNCATION_NOTE;
-  }
-  return out;
-}
-
-/**
- * THE FOUR DEFANGING STEPS, in one place. `sanitizeBody` and `defangInline` each
- * carried a verbatim copy of the chain, so the numbered contract on
- * `sanitizeBody` described one of them and was true of the other only by
- * inspection -- two copies held together by prose. Steps 1-4 of that list ARE
- * this function; each caller owns only what it adds afterwards.
- *
- * @param {unknown} text
- */
-function defangDangerous(text) {
-  return String(text ?? '')
-    .replace(HTML_COMMENT_RE, '')
-    .replace(DANGLING_COMMENT_OPEN_RE, '<!‑-')
-    .replace(MARKER_TOKEN_RE, defangToken)
-    .replace(/@(?=[A-Za-z0-9])/g, '@​');
-}
-
-/**
- * The defanging every model-or-PR-controlled short string gets before a poster
- * may render it, plus the whitespace collapse a one-line field needs. Length
- * policy is NOT here -- it belongs to the caller, because a `class` label and a
- * file path have opposite needs (a label is a tag to cap hard; a path is an
- * identity to keep whole).
- */
-function defangInline(text) {
-  return defangDangerous(text).replace(/\s+/g, ' ').trim();
-}
-
-/**
- * A short label, held to a tighter budget than a body. Same defanging: `class` is
- * model-controlled and a poster that renders it into the comment would carry a
- * marker just as well as `body` would.
- *
- * @param {unknown} text
- */
-export function sanitizeLabel(text) {
-  return defangInline(text).slice(0, MAX_CLASS_CHARS);
-}
-
-/**
- * A file path budget. Borrowing `class`'s 60-char cap here truncated
- * `.../property/property-table.tsx` to a name that exists nowhere, and let two
- * sibling files render as the SAME string -- 1,251 of 6,633 tracked paths
- * exceed 60 chars. The longest tracked path is 188 bytes, so at 500 the cap is
- * unreachable for any real path and exists only against a hostile one padding
- * the posted summary.
- */
-const MAX_PATH_CHARS = 500;
-
-/**
- * A path a poster will render: defanged like a label, but kept WHOLE -- its
- * entire job is naming a real file the reviewer never read. If a hostile path
- * does exceed the cap, the truncation says so and stays unambiguous: the tail
- * carries the cut length and a digest of the full sanitised string, so
- * TRUNCATION cannot collapse two paths the way the 60-char slice made
- * `property-table.tsx` and `property-header.tsx` collapse into one.
- *
- * That is the whole guarantee, and it is narrower than "distinct paths always
- * render distinctly". DEFANGING is lossy and runs BEFORE the digest, so paths
- * differing only in what defanging removes still collide -- measured, all legal
- * git paths: `dir/a<!--x-->b.ts` vs `dir/ab.ts`; two spaces vs one; a tab vs a
- * space; a leading space vs none; and `ifc-lite-review.ts` vs its U+2011
- * non-breaking-hyphen lookalike -- IN ANY CASE now, not only in lowercase. The
- * replacement used to be a fixed lowercase string, which spared the uppercase
- * pair by rewriting its case; that was a worse bug than the collision it
- * avoided, so `defangToken` preserves the match and the pair collides like every
- * other. Sub-cap paths get no digest at all, so nothing disambiguates them. A
- * reader can therefore still see `omitted=2` above two identical-looking
- * entries.
- *
- * Accepted rather than fixed: defanging is load-bearing (it is what stops a
- * path forging a marker) and it must stay lossy to do that job. The cost is
- * cosmetic -- a duplicate-looking line in an advisory list.
- *
- * Digesting the RAW path instead would in fact disambiguate these, and safely:
- * a sha256 hex digest is `[0-9a-f]` only, so it cannot reproduce `<!--`, the
- * marker token or an @-mention. An earlier version of this comment claimed
- * otherwise and was simply wrong. The real cost is that it would hang a digest
- * suffix on EVERY row to disambiguate a case nobody has hit, which is a poor
- * trade for an advisory list.
- *
- * @param {unknown} text
- */
-export function sanitizePath(text) {
-  const out = defangInline(text);
-  if (out.length <= MAX_PATH_CHARS) return out;
-  const digest = createHash('sha256').update(out).digest('hex').slice(0, 12);
-  return `${out.slice(0, MAX_PATH_CHARS)} [truncated: ${out.length} chars, sha256 ${digest}]`;
-}
-
-/**
- * PROOF OF WORK. The anti-#1644 check, and the only one here that a model cannot
- * satisfy by guessing.
- *
- * Set equality in BOTH directions. A MISSING file is the quiet quit. An EXTRA file
- * is a model reporting on something it was never given, which is at least as
- * alarming and would otherwise pass a subset check. Duplicates in
- * `files_reviewed` collapse into the set, which is why the input builder is
- * required to de-duplicate `files` -- otherwise the two sides could differ in
- * multiplicity and this check would not see it.
- *
- * `unreviewable` paths are absent from `files`, so naming one here fails as an
- * extra. That is correct: those files were deliberately not sent, so a review of
- * one is a review of something the model invented.
- */
-function checkProofOfWork({ response, input }) {
-  const expected = new Set(input.files.keys());
-  const claimed = new Set(response.files_reviewed);
-  const missing = [...expected].filter((p) => !claimed.has(p));
-  const extra = [...claimed].filter((p) => !expected.has(p));
-  if (missing.length > 0 || extra.length > 0) {
-    throw new ValidateFindingsError(
-      'PROOF_OF_WORK_FAILED',
-      '`files_reviewed` is not the set of files that were sent.' +
-        (missing.length > 0 ? ` NOT REVIEWED: ${missing.join(', ')}.` : '') +
-        (extra.length > 0 ? ` NEVER SENT: ${extra.join(', ')}.` : '') +
-        ' A model that stopped early cannot report on files it never opened, which is exactly what ' +
-        'claude-code-action#1644 does while exiting 0. REMEDY: re-run; if it recurs, read `num_turns` ' +
-        'in the review step\'s log rather than re-running indefinitely.',
-    );
-  }
-
-  const rc = response.riskiest_change;
-  const file = input.files.get(rc.path);
-  if (!file) {
-    throw new ValidateFindingsError(
-      'PROOF_OF_WORK_FAILED',
-      `\`riskiest_change.path\` is \`${rc.path}\`, which was never sent. REMEDY: re-run.`,
-    );
-  }
-  if (!quoteAppearsIn(file.patch, rc.quoted_line, MIN_PROOF_QUOTE_CHARS)) {
-    throw new ValidateFindingsError(
-      'PROOF_OF_WORK_FAILED',
-      `\`riskiest_change.quoted_line\` is not a line of \`${rc.path}\`'s patch (or is shorter than ` +
-        `${MIN_PROOF_QUOTE_CHARS} characters, which would not be evidence of anything): ` +
-        `${JSON.stringify(String(rc.quoted_line).slice(0, 120))}. This is the one thing a model that ` +
-        'quit early cannot fake. REMEDY: re-run. Quote a WHOLE line, not a fragment; and if the ' +
-        'line you nominated is too long to reproduce exactly, nominate a SHORTER line from the ' +
-        'same file instead -- any real line of the diff proves you read it.',
-    );
-  }
-}
-
-/** The top-level shape. An individual finding is validated -- and dropped -- later. */
-function checkSchema(response) {
-  const fail = (msg) => {
-    throw new ValidateFindingsError('SCHEMA_INVALID', `${msg} REMEDY: fix the prompt's output contract.`);
-  };
-  if (response.verdict !== 'clean' && response.verdict !== 'findings') {
-    fail(`\`verdict\` must be "clean" or "findings"; got ${JSON.stringify(response.verdict)}.`);
-  }
-  if (!Array.isArray(response.files_reviewed) || response.files_reviewed.some((p) => !isNonEmptyString(p))) {
-    fail('`files_reviewed` must be an array of non-empty strings.');
-  }
-  const rc = response.riskiest_change;
-  if (rc === null || typeof rc !== 'object' || Array.isArray(rc) || !isNonEmptyString(rc.path) || !isNonEmptyString(rc.quoted_line)) {
-    // REQUIRED EVEN WHEN CLEAN, and especially then: a clean verdict has no
-    // findings to prove the work with, so this is the ONLY evidence that the model
-    // read anything. Making it optional on `clean` would put the proof exactly
-    // where it is least needed and remove it exactly where it is most.
-    fail('`riskiest_change` must be an object with non-empty `path` and `quoted_line` strings.');
-  }
-  if (!Array.isArray(response.findings)) {
-    fail('`findings` must be an array (empty on a clean verdict, never omitted).');
-  }
-  if (response.verdict === 'clean' && response.findings.length > 0) {
-    throw new ValidateFindingsError(
-      'VERDICT_CONTRADICTS_FINDINGS',
-      `\`verdict\` is "clean" but ${response.findings.length} finding(s) were emitted. Both ways of ` +
-        'resolving this are wrong: trusting the verdict throws away real findings, trusting the ' +
-        'findings posts them under a marker that says the diff was clean. REMEDY: re-run. Never guess ' +
-        'which half was meant.',
-    );
-  }
-}
-
-/**
- * A cross-file claim is only admissible if the harness put the evidence there.
- *
- * The largest defect family in this repository is "the same fix applied at one
- * site when there are two", and until the context pack the reviewer could not
- * see the second site at all. Now it can -- but a model that has been TOLD
- * about a sibling can also invent one, and a fabricated cross-file claim is
- * worse than silence: it sends the author to a file that is fine.
- *
- * So the sibling is checked the same way the anchor quote is: against text the
- * harness retrieved, never against the model's word for it. `path` and `line`
- * must match an excerpt actually placed in the pack, and the quote must appear
- * in that excerpt. A finding whose sibling does not verify is dropped as
- * fabricated, exactly like a bad anchor.
- */
-export function siblingVerifies(sibling, contextPack) {
-  if (sibling == null) return { ok: true, reason: null };       // absent is fine
-  if (typeof sibling !== 'object' || Array.isArray(sibling)) {
-    return { ok: false, reason: '`sibling` is not an object' };
-  }
-  const { path, line, quote } = sibling;
-  if (!isNonEmptyString(path)) return { ok: false, reason: '`sibling.path` is missing' };
-  if (!Number.isInteger(line) || line < 1) return { ok: false, reason: '`sibling.line` is not a line number' };
-  const excerpts = contextPack?.siblings ?? [];
-  if (excerpts.length === 0) {
-    return { ok: false, reason: 'no sibling excerpts were provided, so a cross-file claim has no evidence' };
-  }
-  const near = excerpts.filter((e) => e.path === path && Math.abs(e.line - line) <= 3);
-  if (near.length === 0) {
-    return { ok: false, reason: `no excerpt from \`${path}\` near line ${line} was in the pack` };
-  }
-  if (isNonEmptyString(quote)) {
-    const needle = quote.trim();
-    // ONE WAY ONLY: the excerpt must contain the quote, never the reverse. The
-    // second direction let a fabricated quote pass by merely CONTAINING a real
-    // excerpt line -- "the importer does cache.set(n, scaled); and then silently
-    // drops the alpha channel" verified against an excerpt of
-    // `cache.set(n, scaled);`, because the invented sentence contains it. That
-    // defeats the whole check: the model can wrap one real line in any amount of
-    // invented prose and the harness certifies the lot.
-    //
-    // The reviewer is shown these excerpts, so quoting FROM one is the only
-    // honest direction. A quote longer than the excerpt is not evidence of
-    // anything the harness put there.
-    if (!near.some((e) => e.text.includes(needle))) {
-      return { ok: false, reason: `\`sibling.quote\` is not in the excerpt from \`${path}\`` };
-    }
-  }
-  return { ok: true, reason: null };
-}
-
-/**
- * Per-finding validation. INVALID FINDINGS ARE DROPPED, NOT FATAL, and that is a
- * deliberate asymmetry: a model that gets three of four right should still deliver
- * the three. Every drop is warned about by name, so a silent drop is impossible.
- *
- * A malformed MEMBER is dropped for the same reason a wrong path is -- it is one
- * finding the model got wrong, not a broken contract. The top-level `findings`
- * being the wrong TYPE is fatal, above, because then there is nothing to iterate.
- *
- * STATED HOLE: five garbage findings and one good one exits 0 with one finding.
- * That is the intended trade. The verdict-level check (VALIDATION_EMPTY) is what
- * stands behind it when NOTHING survives.
- */
-
-function validateFindings({ response, input, warn }) {
-  const kept = [];
-  for (const [i, f] of response.findings.entries()) {
-    const drop = (why) => {
-      warn(`DROPPED findings[${i}]: ${why}`);
-      return true;
-    };
-    if (f === null || typeof f !== 'object' || Array.isArray(f)) {
-      drop('not an object.');
-      continue;
-    }
-    if (!isNonEmptyString(f.path)) {
-      drop('`path` is missing or not a non-empty string.');
-      continue;
-    }
-    const sib = siblingVerifies(f.sibling, input.contextPack);
-    if (!sib.ok) {
-      drop(`\`sibling\` does not verify: ${sib.reason}. A cross-file claim the harness cannot confirm is fabricated.`);
-      continue;
-    }
-    const file = input.files.get(f.path);
-    if (!file) {
-      drop(`\`${f.path}\` was never sent to the model, so this finding is about code we did not review.`);
-      continue;
-    }
-    if (typeof f.quote !== 'string' || f.quote.trim().length < MIN_FINDING_QUOTE_CHARS) {
-      drop(
-        `\`quote\` is missing or under ${MIN_FINDING_QUOTE_CHARS} characters, which would not be ` +
-          `evidence of anything: ${JSON.stringify(String(f.quote).slice(0, 120))}.`,
-      );
-      continue;
-    }
-    if (!lineIsAdded(f.line, file.addedLineRanges)) {
-      drop(
-        `\`line\` ${JSON.stringify(f.line)} is not inside an added range of \`${f.path}\` ` +
-          `(${JSON.stringify(file.addedLineRanges)}). Commenting there would annotate code this PR ` +
-          'did not touch.',
-      );
-      continue;
-    }
-    // THE COUPLING CHECK (#3658). `quote` and `line` used to be validated
-    // independently, so a real quote anchored at the wrong added line passed
-    // both checks and posted on the wrong line. Requiring the quote to BE the
-    // text at that exact line closes it. On a mismatch, name where the quote
-    // WAS found (if anywhere) so the drop is diagnosable rather than a bare
-    // refusal -- this is the loud failure direction: the finding is dropped,
-    // never silently re-anchored to a line the model did not name.
-    const matches = addedLinesMatching(file.patch, f.quote);
-    if (!matches.includes(f.line)) {
-      drop(
-        matches.length > 0
-          ? `\`quote\` is not the text at \`line\` ${f.line} of \`${f.path}\` -- it IS the text of added ` +
-            `line(s) ${matches.join(', ')} instead. Posting at ${f.line} would anchor a real finding to ` +
-            'the wrong line, which is worse than not posting it.'
-          : `\`quote\` is not the text of any added line of \`${f.path}\`: ` +
-            `${JSON.stringify(f.quote.slice(0, 120))}.`,
-      );
-      continue;
-    }
-    if (!isNonEmptyString(f.body)) {
-      // An empty body posts a comment that says nothing while the marker counts it
-      // as a finding -- the "marker with an empty body" hole check-review-posted
-      // states about itself. Closed here, where the body still exists.
-      drop('`body` is missing or empty; it would post a comment that says nothing.');
-      continue;
-    }
-    kept.push(f);
-  }
-  return kept;
-}
-
-/**
- * The whole policy, pure over already-read inputs so the harness can drive every
- * branch without touching a filesystem.
- *
- * @returns {{ verdict: string, findings: object[], warnings: string[], counts: object }}
- */
-export function validate({ response, input, onWarn = null }) {
-  const warnings = [];
-  // WARNINGS ARE EMITTED AS THEY HAPPEN, not collected and printed by the caller
-  // afterwards. VALIDATION_EMPTY's remedy is "read the DROPPED warnings above",
-  // and on that path `validate` THROWS -- so a caller that printed the returned
-  // array would print nothing at all, and the remedy would point at output that
-  // does not exist. A gate whose remedy contradicts its finding is worse than one
-  // with no remedy. Caught by its own test, which is why the sink is a parameter.
-  const warn = (m) => {
-    warnings.push(m);
-    if (onWarn) onWarn(m);
-  };
-
-  // THE SENTINEL FIRST, before the field-by-field schema pass. A response that
-  // stopped early usually fails several schema checks at once, and reporting the
-  // first missing field would send the reader to fix the prompt when the real
-  // problem is the token budget. The sentinel names the actual cause.
-  if (response.end !== SENTINEL) {
-    throw new ValidateFindingsError(
-      'RESPONSE_TRUNCATED',
-      `The terminal sentinel is ${JSON.stringify(response.end)}, not ${JSON.stringify(SENTINEL)}. ` +
-        'Valid JSON is not evidence of a complete response: `{"verdict":"clean"}` parses perfectly ' +
-        'and reviewed nothing. The sentinel is the LAST field the model writes, so its absence means ' +
-        'the response ended before the model meant it to. REMEDY: raise the output token budget, or ' +
-        'send fewer files per run.',
-    );
-  }
-
-  checkSchema(response);
-  checkProofOfWork({ response, input });
-
-  let kept = validateFindings({ response, input, warn });
-  const survived = kept.length;
-
-  if (response.verdict === 'findings' && survived === 0) {
-    throw new ValidateFindingsError(
-      'VALIDATION_EMPTY',
-      `The model reported ${response.findings.length} finding(s) and NONE survived validation. ` +
-        'Not downgraded to clean, which would post a verdict the model never gave, and not passed ' +
-        'through empty, which would leave the marker claiming findings that do not exist. ' +
-        'REMEDY: read the DROPPED warnings above -- they name what was wrong with each one.',
-    );
-  }
-
-  let capped = 0;
-  if (kept.length > MAX_FINDINGS) {
-    capped = kept.length - MAX_FINDINGS;
-    warn(
-      `CAPPED: ${kept.length} valid findings, keeping the first ${MAX_FINDINGS} in the model's own ` +
-        `order and dropping ${capped}. That order is not a severity order (stated hole 5).`,
-    );
-    kept = kept.slice(0, MAX_FINDINGS);
-  }
-
-  // SANITISED LAST. Every check above compared against the RAW model text, so
-  // "verbatim" meant verbatim; defanging first would have made a quote of a line
-  // containing the marker token fail its own verbatim check and be dropped as a
-  // fabrication, which is the wrong diagnosis and the wrong remedy.
-  const findings = kept.map((f) => ({
-    path: f.path,
-    line: f.line,
-    quote: sanitizeBody(f.quote),
-    // Sanitised FIRST, then required non-empty. Checking the raw body let a
-    // finding whose body is only an HTML comment pass validation, sanitise to the
-    // empty string, and be refused downstream by post-review as BAD_FINDING -- a
-    // red job with no marker, for input this validator had certified. Two files
-    // in one change disagreeing about the same contract.
-    body: sanitizeBody(f.body),
-    class: sanitizeLabel(f.class ?? 'unclassified') || 'unclassified',
-    // CARRIED THROUGH, because verifying it and then dropping it is worse than
-    // never checking. `siblingVerifies` above proves the excerpt the finding
-    // names is really in the pack at the line it claims -- and this map then
-    // emitted everything BUT the sibling, so the judge read "verified sibling:
-    // none" on every finding and post-review could not render one either. The
-    // defect family this repository calls its largest -- a fix applied at one of
-    // two sites -- reached the judge stripped of the single piece of evidence
-    // supporting it, beside a rubric that says to drop assertions the quoted
-    // lines do not show. It was the class most likely to be deleted, and the
-    // deletion is not fail-soft.
-    ...(f.sibling
-      ? {
-          sibling: {
-            path: f.sibling.path,
-            line: f.sibling.line,
-            ...(f.sibling.quote ? { quote: sanitizeBody(f.sibling.quote) } : {}),
-          },
-        }
-      : {}),
-  }));
-
-  // A finding whose body sanitises to nothing is DROPPED here rather than
-  // certified. It would otherwise reach post-review, which refuses an empty body
-  // as BAD_FINDING and reddens the job with no marker -- for input this file had
-  // just approved.
-  const nonEmpty = findings.filter((f) => f.body.trim() !== '');
-  if (findings.length > 0 && nonEmpty.length === 0) {
-    throw new ValidateFindingsError(
-      'VALIDATION_EMPTY',
-      'Every surviving finding sanitised to an empty body. Reporting `clean` here would be a lie ' +
-        'and reporting findings would name comments that cannot be posted. REMEDY: re-run.',
-    );
-  }
-
-
-  return {
-    verdict: response.verdict,
-    findings: nonEmpty,
-    warnings,
-    counts: { emitted: response.findings.length, surviving: survived, capped, kept: findings.length },
-  };
-}
-
-/**
- * The paths the reviewer NEVER SAW THE CONTENT OF, ready for a posted comment
- * body. Both ways that happens count: dropped to fit the model prompt (#3679),
- * and refused a patch by GitHub for being too large.
- *
- * IT USED TO BE ONLY THE FIRST, matched by `reason === OMITTED_FOR_PROMPT_REASON`
- * inline, HERE, while build-review-input.mjs's own PARTIAL REVIEW log matched
- * a DIFFERENT inline predicate over the same rows. A PR whose one unreviewable
- * file was one GitHub declined to send therefore produced a marker
- * byte-identical to a full review's, and CodeRabbit stood down on it -- the
- * absence-reads-as-success shape one layer below where #3679 put the
- * disclosure. Both call sites now share `isUnread`, exported from
- * build-review-input.mjs next to the `kind` field it reads, so the log line
- * and the posted marker cannot drift apart the way two independently spelled
- * copies did.
- *
- * A deletion or a pure rename is NOT counted: there was no changed content for
- * the reviewer to read, so nothing is being withheld and disclosing it would
- * train readers to ignore the warning.
- *
- * Sanitised HERE, because this file is the boundary between model-or-PR
- * controlled bytes and the poster: a git path may contain any byte but NUL and
- * `/`, including `-->` and the literal marker token, and an unsanitised path in
- * the summary would be a marker-forgery channel opened by the very feature
- * whose job is to keep absence visible. `sanitizePath` defangs the token and
- * strips HTML comments but keeps the path WHOLE (`sanitizeLabel`'s 60-char
- * `class` cap stood here once and rewrote real paths into names that exist
- * nowhere -- and collapsed sibling files into one string); a path that
- * sanitises to nothing still has to appear, so it is named as unprintable
- * rather than dropped.
- *
- * @param {{path: string, reason?: string}[]} unreviewable
- * @returns {string[]}
- */
-export function omittedForPromptPaths(unreviewable) {
-  return unreviewable
-    .filter(isUnread)
-    .map((u) => sanitizePath(u.path) || '(a path that sanitised to nothing)');
-}
+export { ValidateFindingsError };
+export { stripFence, readInput };
+export { SENTINEL, MAX_FINDINGS, validate, omittedForPromptPaths };
+export { MAX_BODY_CHARS, sanitizeBody, sanitizeLabel, sanitizePath };
+export { siblingVerifies };
 
 function main() {
   const args = parseArgs(process.argv.slice(2));
